stage-0/gcc: use our own sh as interpreter
Unfortunately, this does not entirely removes the requirement of
a /bin/sh being present, as awk(1) is invoked during the build to
use popen(3) in one place, a libc function that has a hard-coded
reference to /bin/sh. Nevertheless, using our interpreter wherever
possible greatly reduces the dependency on the host system.