CC toolchains: remove wrong reference to $(TOOLCHAIN) in PATH
The environment variable PATH is taken literally by the shell, without
any expansion. So, having a string like "$(TOOLCHAIN)/clang/bin"
in PATH does not change anything as we do not have a directory
with that name. Drop it for now. In the future we might add "bin
dirs" to the CC toolchain as well, but as long as our rules do not
support this, drop the wrong entry here.