etc/repos.json: have file root have the to_git pragma
The main use of the bootstrappable toolchain is to just-import-git(1)
it as toolchain to an existing project. However, it is an explicitly
supported use case to build just the toolchain as stand-alone binaries.
Therefore, also for that use case, make sure all repositories are
content fixed, so that the various bootstrap stages are eligible
for caching as export targets.